Fix launching the example app on new archotecture
