+++ modules/react-native-wallet/android/src/oldarch/WalletSpec.kt
@@ -12,5 +12,5 @@ abstract class WalletSpec internal constructor(context: ReactApplicationContext)
   abstract fun getSecureWalletInfo(promise: Promise)
   abstract fun getCardStatus(last4Digits: String, promise: Promise)
   abstract fun getCardTokenStatus(tsp: String, tokenRefId: String, promise: Promise)
-  abstract fun addCardToWallet(data: ReadableMap, promise: Promise)
+  abstract fun addCardToWallet(cardData: ReadableMap, promise: Promise)
 }
