+++ modules/react-native-wallet/example/android/gradle.properties
@@ -40,8 +40,8 @@ reactNativeArchitectures=armeabi-v7a,arm64-v8a,x86,x86_64
 # to write custom TurboModules/Fabric components OR use libraries that
 # are providing them.
 # Note that this is incompatible with web debugging.
-#newArchEnabled=true
-#bridgelessEnabled=true
+newArchEnabled=true
+bridgelessEnabled=true
 
 # Uncomment the line below to build React Native from source.
 #react.buildFromSource=true
