--- modules/react-native-wallet/src/types.ts
@@ -1,50 +1,0 @@
-type WalletData = AndroidWalletData | IOSWalletData;
-
-type AndroidWalletData = {
-  platform: 'android';
-  deviceID: string;
-  walletAccountID: string;
-};
-
-type IOSWalletData = {
-  platform: 'ios';
-  nonce: string;
-  nonceSignature: string;
-  certificates: string;
-};
-
-type UserAddress = {
-  name: string;
-  addressOne: string;
-  addressTwo?: string;
-  administrativeArea: string;
-  locality: string;
-  countryCode: string;
-  postalCode: string;
-  phoneNumber: string;
-};
-
-type AndroidCardData = {
-  platform: 'android';
-  network: string;
-  opaquePaymentCard: string;
-  cardHolderName: string;
-  lastDigits: string;
-  userAddress: UserAddress;
-};
-
-type CardStatus =
-  | 'not found'
-  | 'requireActivation'
-  | 'activating'
-  | 'activated'
-  | 'suspended'
-  | 'deactivated';
-
-export type {
-  WalletData,
-  AndroidWalletData,
-  UserAddress,
-  AndroidCardData,
-  CardStatus,
-};
