+++ modules/react-native-wallet/src/utils.ts
@@ -1,4 +1,4 @@
-import type { CardStatus } from './types';
+import type { CardStatus } from './NativeWallet';
 
 function getCardState(stateId: number): CardStatus {
   switch (stateId) {
